requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-groq
google-genai
sqlalchemy
psycopg2-binary
python-dotenv